+++ frontend/src/features/mindmap/types/mindmap.ts
@@ -1,33 +1,22 @@
-export type NodeId = string;
+export type MindmapType = "ALL" | "PUBLIC" | "PRIVATE";
 
-export type Node = {
-    id: NodeId;
-
-    x: number;
-    y: number;
-
-    width: number;
-    height: number;
-
-    data: NodeData;
+export type MindmapItem = {
+    mindmapId: string;
+    mindmapName: string;
+    createdAt: string;
+    updatedAt: string;
+    isFavorite: boolean;
 };
 
-export type NodeType = "root" | "normal";
+export type ActivityCategory = "INTERN" | "STUDY" | "CLUB" | "PROJECT" | "VOLUNTEER" | "PARTTIME" | "CONTEST" | "ETC";
 
-export type NodeElement = Node & {
-    parentId: NodeId;
-
-    // double linked list
-    nextId: NodeId | null;
-    prevId: NodeId | null;
-
-    firstChildId: NodeId | null;
-    lastChildId: NodeId | null;
-
-    type: NodeType;
-};
-
-export type NodeData = {
-    contents: string;
-    pakxepakxe?: "뭔 타입이 올지 모르겠으니..";
-};
+export const ACTIVITY_CATEGORIES: ReadonlyArray<{ id: ActivityCategory; label: string; emoji: string }> = [
+    { id: "INTERN", label: "인턴", emoji: "💼" },
+    { id: "STUDY", label: "학업", emoji: "📚" },
+    { id: "CLUB", label: "동아리", emoji: "🎯" },
+    { id: "PROJECT", label: "프로젝트", emoji: "🚀" },
+    { id: "PARTTIME", label: "아르바이트", emoji: "💰" },
+    { id: "VOLUNTEER", label: "봉사활동", emoji: "🍀" },
+    { id: "CONTEST", label: "공모전", emoji: "🏆" },
+    { id: "ETC", label: "기타", emoji: "✨" },
+];
